# swagger.yaml
openapi: 3.0.3
info:
  title: SenAI CRM API
  description: Agentic CRM Intelligence Platform
  version: 1.0.0
  contact:
    name: SenAI CRM Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
paths:
...
  /contacts/{email}:
    get:
      summary: Get contact by email
      operationId: getContact
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Contact profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactOut'
        '404':
          $ref: '#/components/responses/NotFound'
    
    patch:
      summary: Update contact fields
      operationId: updateContact
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                tier:
                  type: string
                account_value:
                  type: number
                vip_status:
                  type: boolean
                churn_risk_score:
                  type: number
      responses:
        '200':
          description: Contact updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  updated:
                    type: boolean
                  fields:
                    type: array
                    items:
                      type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /emails/:
    get:
      summary: List all emails with filters
      operationId: listEmails
      tags: [Emails]
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: category
          in: query
          schema:
            type: string
        - name: urgency
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: requires_human
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /emails/{email_id}:
    get:
      summary: Get single email with full details
      operationId: getEmail
      tags: [Emails]
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email details
        '404':
          $ref: '#/components/responses/NotFound'
  /emails/{email_id}/reasoning:
    get:
      summary: Get AI reasoning trace for an email
      operationId: getEmailReasoning
      tags: [Emails]
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reasoning trace
        '404':
          $ref: '#/components/responses/NotFound'
  /emails/{email_id}/rag-context:
    get:
      summary: Get RAG chunks used for this email
      operationId: getEmailRAGContext
      tags: [Emails]
      parameters:
        - name: email_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: RAG context chunks
        '404':
          $ref: '#/components/responses/NotFound'
  /emails/bulk/spam:
    post:
      summary: Bulk mark emails as spam
      operationId: bulkMarkSpam
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Bulk update result
  /emails/bulk/assign:
    post:
      summary: Bulk assign emails to team member
      operationId: bulkAssign
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_ids:
                  type: array
                  items:
                    type: string
                assignee:
                  type: string
      responses:
        '200':
          description: Bulk assign result
  /emails/bulk/archive:
    post:
      summary: Bulk archive emails
      operationId: bulkArchive
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Bulk archive result
...
components:
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error_code
        - message
        - details
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
...
    ContactOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        tier:
          type: string
        account_value:
          type: number
        vip_status:
          type: boolean
        churn_risk_score:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error_code: NOT_FOUND
            message: Resource not found
            details: {}
...